+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.5 - (2022-05-01)
+
+* Function to collapse a list of exceptions to raise a MultipleError.
+* Use local script logger for logging.
+
 ## 0.1.4 - (2022-03-07)
 
 * Fix name: `ExpiredAPIBaseError` -> `ExpiredSessionError`